requirements.txt
Django>=4.0,<=5.13
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
gunicorn>=21.0.0
whitenoise>=6.5.0
